refactor: split publishers in different files for better readability
